chore(asm): refactor out django appsec urls
