+++ tests/contrib/django/views.py
@@ -3,7 +3,6 @@ Class based views used for Django tests.
 """
 
 from functools import partial
-import hashlib
 
 from django.contrib.auth.models import User
 from django.contrib.syndication.views import Feed
@@ -17,9 +16,6 @@ from django.views.generic import TemplateView
 from django.views.generic import View
 
 from ddtrace import tracer
-from ddtrace.appsec import _asm_request_context
-from ddtrace.appsec.iast._util import _is_python_version_supported as python_supported_by_iast
-from ddtrace.appsec.trace_utils import block_request_if_user_blocked
 from ddtrace.contrib.trace_utils import set_user
 
 
@@ -185,10 +181,6 @@ def not_found_view(request):
     raise Http404("DNE")
 
 
-def path_params_view(request, year, month):
-    return HttpResponse(status=200)
-
-
 def identify(request):
     set_user(
         tracer,
@@ -200,71 +192,3 @@ def identify(request):
         scope="usr.scope",
     )
     return HttpResponse(status=200)
-
-
-def body_view(request):
-    # Django >= 3
-    if hasattr(request, "headers"):
-        content_type = request.headers["Content-Type"]
-    else:
-        # Django < 3
-        content_type = request.META["CONTENT_TYPE"]
-    if content_type in ("application/json", "application/xml", "text/xml"):
-        data = request.body
-        return HttpResponse(data, status=200)
-    else:
-        data = request.POST
-        return HttpResponse(str(dict(data)), status=200)
-
-
-def weak_hash_view(request):
-    m = hashlib.md5()
-    m.update(b"Nobody inspects")
-    m.update(b" the spammish repetition")
-    m.digest()
-    return HttpResponse("OK", status=200)
-
-
-def block_callable_view(request):
-    _asm_request_context.block_request()
-    return HttpResponse("OK", status=200)
-
-
-def checkuser_view(request, user_id):
-    block_request_if_user_blocked(tracer, user_id)
-    return HttpResponse(status=200)
-
-
-def taint_checking_enabled_view(request):
-    if python_supported_by_iast():
-        from ddtrace.appsec.iast._taint_tracking import is_pyobject_tainted
-    else:
-
-        def is_pyobject_tainted(x):
-            return True
-
-    # TODO: Taint request body
-    # assert is_pyobject_tainted(request.body)
-    assert is_pyobject_tainted(request.GET["q"])
-    assert is_pyobject_tainted(request.META["QUERY_STRING"])
-    assert is_pyobject_tainted(request.META["HTTP_USER_AGENT"])
-    assert is_pyobject_tainted(request.headers["User-Agent"])
-
-    return HttpResponse(request.META["HTTP_USER_AGENT"], status=200)
-
-
-def taint_checking_disabled_view(request):
-    if python_supported_by_iast():
-        from ddtrace.appsec.iast._taint_tracking import is_pyobject_tainted
-    else:
-
-        def is_pyobject_tainted(x):
-            return False
-
-    assert not is_pyobject_tainted(request.body)
-    assert not is_pyobject_tainted(request.GET["q"])
-    assert not is_pyobject_tainted(request.META["QUERY_STRING"])
-    assert not is_pyobject_tainted(request.META["HTTP_USER_AGENT"])
-    assert not is_pyobject_tainted(request.headers["User-Agent"])
-
-    return HttpResponse(request.META["HTTP_USER_AGENT"], status=200)
